Split cli docs and add basic docs run
